Update README.md with shell of project structure
Add in planned project structure. To Professor Wloka: I did not
initially use a github repository for this project, so I have created
one here and will add in my current progress. I created the README
first in order to create a shell with which to track my progress and
organize my work as I go.