Finish field lines field_lines_detector script
I realized that there was no ground truth of field lines, and as such
there was no way to accruately and automatically hough transform
every image via the field lines. However, upong detecting the yard lines
I noticed that they could be used to remove a lot of the background
and sidelines. Again, with no ground truth of where the field is in
the image comapred to the sidelines, there was no way to do this with
model. However, with the logic of:
The yard lines are always visible ->
The yard lines only appear on the field ->
We can trim the area of the image outside of the yard lines ->
We can trim the parts of the image that are not mostly green ->
We can trim most of the background and sidelines without training a
model,

I was able to create a pre processing script that will eliminate most
noise from images.